+++ configs/lpc4357-evb/src/lpc43_buttons.c
@@ -129,10 +129,10 @@ void board_button_initialize(void)
  *
  ****************************************************************************/
 
-uint8_t board_buttons(void)
+uint32_t board_buttons(void)
 {
 #if 0 /* Not yet implemented */
-  uint8_t ret = 0;
+  uint32_t ret = 0;
   int i;
 
   /* Check that state of each key */
