+++ configs/olimex-efm32g880f128-stk/src/efm32_buttons.c
@@ -125,9 +125,9 @@ void board_button_initialize(void)
  *
  ****************************************************************************/
 
-uint8_t board_buttons(void)
+uint32_t board_buttons(void)
 {
-  uint8_t ret;
+  uint32_t ret;
   int i;
 
   /* Check each button */
