+++ configs/sam3u-ek/src/sam_buttons.c
@@ -133,9 +133,9 @@ void board_button_initialize(void)
  *
  ************************************************************************************/
 
-uint8_t board_buttons(void)
+uint32_t board_buttons(void)
 {
-  uint8_t retval;
+  uint32_t retval;
 
   retval  = sam_gpioread(GPIO_BUTTON1) ? 0 : BUTTON1;
   retval |= sam_gpioread(GPIO_BUTTON2) ? 0 : BUTTON2;
