+++ configs/sama5d2-xult/src/sam_buttons.c
@@ -96,7 +96,7 @@ void board_button_initialize(void)
  *
  ****************************************************************************/
 
-uint8_t board_buttons(void)
+uint32_t board_buttons(void)
 {
   return sam_pioread(PIO_BTN_USER) ? 0 : BUTTON_USER_BIT;
 }
