+++ configs/stm32vldiscovery/src/stm32_buttons.c
@@ -75,9 +75,9 @@ void board_button_initialize(void)
  * Name: board_buttons
  ****************************************************************************/
 
-uint8_t board_buttons(void)
+uint32_t board_buttons(void)
 {
-  uint8_t ret = 0;
+  uint32_t ret = 0;
 
   ret = (stm32_gpioread(g_buttons[i]) == false ? 1 : 0);
